% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{generate.data}
\alias{generate.data}
\title{Generate synthetic data for simulation study}
\usage{
generate.data(
  dimensions = 2,
  w1 = 0.6,
  n1 = 40,
  n2 = 40,
  n3 = 10,
  n4 = 10,
  m = 500,
  mu1 = c(-0.25, -0.25),
  mu2 = c(0.25, 0.25),
  mu3 = c(0.75, -0.75),
  mu4 = c(-0.75, 0.75),
  sigma1 = 0.02,
  sigma2 = 0.02,
  sigma3 = 0.01,
  sigma4 = 0.01,
  theta = 1.5,
  beta = 10,
  seed = NULL
)
}
\arguments{
\item{dimensions}{}

\item{seed}{}
}
\value{
An object of class \code{l1synthetic} with the following elements:
\item{votes}{}
\item{legis_data}{}
\item{votes_data}{}
\item{true_wight}{}
\item{true_beta}{}
\item{code}{}
}
\description{
Generate synthetic data for simulation study
}
\references{
Sooahn Shin, Yohan Lim, and Jong Hee Park 2019. "L1 norm Based Multidimensional Ideal Point Estimation: With Application to Roll Call Voting Data" Working Paper.
}
