% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1ideal.R
\name{l1ideal}
\alias{l1ideal}
\title{L1 norm based multidimensional ideal point estimation}
\usage{
l1ideal(rollcall, dimensions = 2, mcmc = 5000, thin = 1,
  burnin = 0, minvotes = 20, lop = 0, polarity = c(1, 1),
  verbose = 100, seed = NULL)
}
\arguments{
\item{rollcall}{A \code{rollcall} object generated by Simon Jackman's \code{pscl} package.}

\item{dimensions}{The number of dimensions in the latent space. This function supports either 1 or 2 number of dimensions. The default is \code{2}.}

\item{mcmc}{The total number of MCMC iterations. The default is \code{5000}.}

\item{thin}{A positive integer indicating the thining interval for the Markov chain. The default is \code{1}.}

\item{burnin}{The burnin interval for the Markov chain. The default is \code{0}.}

\item{minvotes}{Minimum number of votes required for a legislator; i.e., legislators who have voted less than \code{minvotes} are excluded from the analysis. The default is \code{20}.}

\item{lop}{Minimum proportion of the minority votes; i.e., votes where the minority side is smaller than \code{lop} percent are excluded from the analysis. Recommend to keep the default value \code{0}, since every vote includes valuable information.}

\item{polarity}{A vector specifying the row index(es) of the legislator(s) in 'votes' matrix constrained to have a positive value on each dimension; e.g., \code{c(10,20)} indicates that the 10th legislator is constrained to have a positive value on the first dimension while the 20th legislator is constrained to have a positive value on the second dimension. The length of the vector must corresponds to the number of dimensions. The default is \code{c(1,1)}.}

\item{verbose}{A positive integer specified to print the progress on the screen; e.g., every \code{verbose}-th iteration is printed on the screen. The default is \code{100}.}

\item{seed}{An integer used to set the seed. The default is \code{NULL}.}
}
\value{
An object of class \code{l1ideal} with the following elements:
\item{beta}{An object of class \code{mcmc} containing the posterior samples of the beta parameter.}
\item{weight}{An object of class \code{mcmc} containing the posterior samples of the weight of the first dimension.}
\item{legislators}{A list of \code{mcmc} object(s) containing the posterior samples of the legislator ideal points. Posterior samples of each dimensional coordinates are stored in a separate \code{mcmc}.}
\item{yea_positions}{A list of \code{mcmc} object(s) containing the posterior samples of the yea positions. Posterior samples of each dimensional coordinates are stored in a separate \code{mcmc}.}
\item{nay_positions}{A list of \code{mcmc} object(s) containing the posterior samples of the nay positions. Posterior samples of each dimensional coordinates are stored in a separate \code{mcmc}.}
\item{legis_data}{A matrix or data frame containing covariates of legislators included in the \code{rollcall} object.}
\item{votes_data}{A matrix or data frame containing covariates of votes included in the \code{rollcall} object.}
\item{legis_dropped}{A vector of excluded legislators.}
\item{votes_dropped}{A vector of excluded votes.}
\item{code}{The full code that has been implemented.}
\item{running_time}{The total running time of the estimation.}
}
\description{
L1 norm based multidimensional ideal point estimation
}
\examples{
\dontrun{
dat <- generate.data(seed = 1)
rc <- pscl::rollcall(dat$votes,
                     yea = 1, nay = 0, missing = 2, notInLegis = 3,
                     legis.names = dat$legis_data$name, legis.data = dat$legis_data,
                     vote.names = dat$votes_data$name, vote.data = dat$votes_data)
res <- l1ideal(rc, dimensions = 2, mcmc  = 100, thin = 10, burnin = 100, minvotes = 20,
               lop = 0, polarity = c(86,97), verbose = 100, seed = 123)
p <- plot.simulation(dat, res)
multiplot(plotlist = p, cols = 4)
}
}
\references{
Sooahn Shin, Yohan Lim, and Jong Hee Park 2019. "L1 norm Based Multidimensional Ideal Point Estimation: With Application to Roll Call Voting Data" Working Paper.
}
