% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{multiplot}
\alias{multiplot}
\title{Multiple ggplots}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{layout}{}
}
\value{

}
\description{
Multiple ggplots
}
\references{
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
