% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{plot.simulation}
\alias{plot.simulation}
\title{Plot the simulation study}
\usage{
\method{plot}{simulation}(synthetic_data, l1object, coord_flip = FALSE)
}
\arguments{
\item{coord_flip}{}
}
\value{

}
\description{
Plot the simulation study
}
\references{
Sooahn Shin, Yohan Lim, and Jong Hee Park 2019. "L1 norm Based Multidimensional Ideal Point Estimation: With Application to Roll Call Voting Data" Working Paper.
}
