% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1ideal.R
\name{postprocess.l1ideal}
\alias{postprocess.l1ideal}
\title{Postprocess the estimates to maximize the posterior}
\usage{
postprocess.l1ideal(l1object)
}
\arguments{
\item{l1object}{An object of class \code{l1ideal}}
}
\value{
An object of class \code{l1ideal} where the estimates for legislator ideal points, yea positions, and nay positions are shifted with the value of \Delta^\ast to maximize the posterior.
}
\description{
Postprocess the estimates to maximize the posterior
}
\references{
Sooahn Shin, Johan Lim, and Jong Hee Park 2019. "L1 norm Based Multidimensional Ideal Point Estimation: With Application to Roll Call Voting Data" Working Paper.
}
