% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l1ideal.R
\name{summary.l1ideal}
\alias{summary.l1ideal}
\title{Summarise the l1ideal object}
\usage{
\method{summary}{l1ideal}(l1object)
}
\arguments{
\item{l1object}{An object of class \code{l1ideal}}
}
\value{
A data.frame including the mean, standard deviation, and credible interval of each parameter.
}
\description{
Summarise the l1ideal object
}
\references{
Sooahn Shin, Yohan Lim, and Jong Hee Park 2019. "L1 norm Based Multidimensional Ideal Point Estimation: With Application to Roll Call Voting Data" Working Paper.
}
