\name{l1ideal-package}
\alias{l1ideal-package}
\alias{l1ideal}
\docType{package}
\title{
\packageTitle{l1ideal}
}
\description{
\packageDescription{l1ideal}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{l1ideal}
\packageIndices{l1ideal}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{l1ideal}

Maintainer: \packageMaintainer{l1ideal}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
